+++ com-mmxlabs-lng-scheduler/com.mmxlabs.scheduler.optmiser/src/main/java/com/mmxlabs/scheduler/optimiser/components/ILoadSlot.java
@@ -30,5 +30,5 @@ public interface ILoadSlot extends IPortSlot {
 	 * 
 	 * @return
 	 */
-	long getPurchasePrice();
+	int getPurchasePrice();
 }
